+++ metagraph/src/annotation/binary_matrix/base/binary_matrix.hpp
@@ -18,7 +18,7 @@ namespace matrix {
 class BinaryMatrix {
   public:
     typedef uint64_t Row;
-    typedef uint64_t Column;
+    typedef uint32_t Column;
 
     typedef SmallVector<Column> SetBitPositions;
     typedef std::function<void(const SetBitPositions &)> RowCallback;
