+++ metagraph/src/annotation/binary_matrix/multi_brwt/clustering.hpp
@@ -25,7 +25,7 @@ struct SparseColumn {
 // `SparseColumn` storing the column size and the positions of its set bits.
 // Output: a set of greedily matched column pairs.
 template <class T>
-std::vector<std::vector<uint64_t>>
+std::vector<std::vector<uint32_t>>
 greedy_matching(const std::vector<T> &columns, size_t num_threads = 1);
 
 // Format resembling the Z matrix from scipy.cluster.hierarchy.linkage
